Move WeakLocks to otter.util.weaklocks
