+++ otter/test/models/test_cass_models.py
@@ -347,36 +347,6 @@ class VerifiedViewTests(SynchronousTestCase):
         self.assertFalse(self.log.msg.called)
 
 
-class WeakLocksTests(SynchronousTestCase):
-    """
-    Tests for `WeakLocks`
-    """
-
-    def setUp(self):
-        """
-        Sample `WeakLocks` object
-        """
-        self.locks = WeakLocks()
-
-    def test_returns_deferlock(self):
-        """
-        `get_lock` returns a `DeferredLock`
-        """
-        self.assertIsInstance(self.locks.get_lock('a'), defer.DeferredLock)
-
-    def test_same_lock(self):
-        """
-        `get_lock` on same uuid returns same `DeferredLock`
-        """
-        self.assertIs(self.locks.get_lock('a'), self.locks.get_lock('a'))
-
-    def test_diff_lock(self):
-        """
-        `get_lock` on different uuid returns different `DeferredLock`
-        """
-        self.assertIsNot(self.locks.get_lock('a'), self.locks.get_lock('b'))
-
-
 scaling_group_entry = {
     'tenantId': 'tenant_id',
     "groupId": 'group_id',
